Refactoring: HttpClientHelper -> HttpProxy
