+++ BusinessAssistant-opendata/src/main/java/com/businessassistantbcn/opendata/service/config/TestService.java
@@ -2,9 +2,10 @@ package com.businessassistantbcn.opendata.service.config;
 
 import com.businessassistantbcn.opendata.config.PropertiesConfig;
 import com.businessassistantbcn.opendata.dto.test.StarWarsVehiclesResultDto;
-import com.businessassistantbcn.opendata.helper.HttpClientHelper;
 import java.net.MalformedURLException;
 import java.net.URL;
+
+import com.businessassistantbcn.opendata.proxy.HttpProxy;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import reactor.core.publisher.Mono;
@@ -16,10 +17,10 @@ public class TestService {
     PropertiesConfig config;
 
     @Autowired
-    HttpClientHelper helper;
+    HttpProxy httpProxy;
 
     public Mono<StarWarsVehiclesResultDto> getTestData() throws MalformedURLException{
-           return helper.getRequestData(new URL(config.getDs_test()), StarWarsVehiclesResultDto.class);
+           return httpProxy.getRequestData(new URL(config.getDs_test()), StarWarsVehiclesResultDto.class);
     }
 
 }
