+++ BusinessAssistant-opendata/src/main/java/com/businessassistantbcn/opendata/service/externaldata/BigMallsService.java
@@ -2,32 +2,22 @@ package com.businessassistantbcn.opendata.service.externaldata;
 
 import java.net.MalformedURLException;
 import java.net.URL;
-import java.util.List;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
+import com.businessassistantbcn.opendata.proxy.HttpProxy;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.cloud.client.circuitbreaker.CircuitBreaker;
 import org.springframework.cloud.client.circuitbreaker.CircuitBreakerFactory;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
-import org.springframework.web.server.ResponseStatusException;
 
 import com.businessassistantbcn.opendata.config.PropertiesConfig;
 import com.businessassistantbcn.opendata.dto.GenericResultDto;
 import com.businessassistantbcn.opendata.dto.bigmalls.BigMallsDto;
 
-import com.businessassistantbcn.opendata.helper.HttpClientHelper;
 import com.businessassistantbcn.opendata.helper.JsonHelper;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.fasterxml.jackson.databind.JavaType;
-import com.fasterxml.jackson.databind.JsonMappingException;
-import com.fasterxml.jackson.databind.ObjectMapper;
 
 import reactor.core.publisher.Mono;
 
@@ -39,7 +29,7 @@ public class BigMallsService {
 	@Autowired
 	private PropertiesConfig config;
 	@Autowired
-	private HttpClientHelper httpClientHelper;
+	private HttpProxy httpProxy;
 	@Autowired
 	private GenericResultDto<BigMallsDto> genericResultDto;
 	@Autowired
@@ -56,7 +46,7 @@ public class BigMallsService {
 			return getPageDefault();
 		}
 
-		Mono<BigMallsDto[]> response = httpClientHelper.getRequestData(url, BigMallsDto[].class);
+		Mono<BigMallsDto[]> response = httpProxy.getRequestData(url, BigMallsDto[].class);
 		CircuitBreaker circuitBreaker = circuitBreakerFactory.create("circuitBreaker");
 
 		return circuitBreaker.run(() -> response.flatMap(dto -> {
